+++ zetaclient/inbound_tracker.go
@@ -92,16 +92,20 @@ func (ob *BitcoinChainClient) CheckReceiptForBtcTxHash(txHash string, vote bool)
 	if err != nil {
 		return "", err
 	}
-	block, err := ob.rpcClient.GetBlockVerbose(blockHash)
+	blockVb, err := ob.rpcClient.GetBlockVerboseTx(blockHash)
 	if err != nil {
 		return "", err
 	}
+	if len(blockVb.Tx) <= 1 {
+		return "", fmt.Errorf("block %d has no transactions", blockVb.Height)
+	}
+	depositorFee := CalcDepositorFee(blockVb, ob.chain.ChainId, ob.netParams, ob.logger.WatchInTx)
 	tss, err := ob.zetaClient.GetBtcTssAddress(ob.chain.ChainId)
 	if err != nil {
 		return "", err
 	}
 	// #nosec G701 always positive
-	event, err := GetBtcEvent(*tx, tss, uint64(block.Height), &ob.logger.WatchInTx, ob.chain.ChainId)
+	event, err := GetBtcEvent(*tx, tss, uint64(blockVb.Height), &ob.logger.WatchInTx, ob.netParams, depositorFee)
 	if err != nil {
 		return "", err
 	}
@@ -117,7 +121,8 @@ func (ob *BitcoinChainClient) CheckReceiptForBtcTxHash(txHash string, vote bool)
 		ob.logger.WatchInTx.Error().Err(err).Msg("error posting to zeta core")
 		return "", err
 	} else if zetaHash != "" {
-		ob.logger.WatchInTx.Info().Msgf("BTC deposit detected and reported: PostVoteInbound zeta tx: %s ballot %s", zetaHash, ballot)
+		ob.logger.WatchInTx.Info().Msgf("BTC deposit detected and reported: PostVoteInbound zeta tx hash: %s inTx %s ballot %s fee %v",
+			zetaHash, txHash, ballot, depositorFee)
 	}
 	return msg.Digest(), nil
 }